requirements.txt
arxiv
openai
python-dotenv
argparse
markdown2
markdown-it-py
mdit-py-plugins 
google-genai
anthropic
tqdm